added check

+++ events/join_event.py
@@ -1,4 +1,3 @@
-from types import NoneType
 import disnake
 from disnake.ext import commands
 import json
@@ -18,46 +17,49 @@ class join_to_create(commands.Cog):
         with open("json/guild.json", "r") as guild_file:
             guild_data = json.load(guild_file)
 
-        if before.channel is None and after.channel is not None:
-            channel = disnake.utils.get(
-                member.guild.voice_channels,
-                id=guild_data[str(member.guild.id)]["join_to_create_channel"]
-            )
-            category = disnake.utils.get(
-                member.guild.categories,
-                id=guild_data[str(member.guild.id)]["join_to_create_category"]
-            )
-            try:
-                if after.channel.id == channel.id:
-                    new_voice = await member.guild.create_voice_channel(
-                        name=f"║🔊・{member.name}'s channel",
-                        category=category,
-                        overwrites={
-                            member: disnake.PermissionOverwrite(
-                                administrator=True
-                            )
-                        }
-                    )
-                    await member.move_to(new_voice)
-                    data["jtcc"][str(new_voice.id)] = {}
-                    data["jtcc"][str(new_voice.id)]["owner"] = member.id
-                    data["jtcc"][str(new_voice.id)]["channel_name"] = new_voice.name
-                    data["jtcc"][str(new_voice.id)]["channel_id"] = new_voice.id
-                    data["jtcc"][str(new_voice.id)]["state"] = "public"
+        if guild_data[str(member.guild.id)]["join_to_create_channel"] and guild_data[str(member.guild.id)]["join_to_create_category"]:
+            if before.channel is None and after.channel is not None:
+                channel = disnake.utils.get(
+                    member.guild.voice_channels,
+                    id=guild_data[str(member.guild.id)]["join_to_create_channel"]
+                )
+                category = disnake.utils.get(
+                    member.guild.categories,
+                    id=guild_data[str(member.guild.id)]["join_to_create_category"]
+                )
+                try:
+                    if after.channel.id == channel.id:
+                        new_voice = await member.guild.create_voice_channel(
+                            name=f"║🔊・{member.name}'s channel",
+                            category=category,
+                            overwrites={
+                                member: disnake.PermissionOverwrite(
+                                    administrator=True
+                                )
+                            }
+                        )
+                        await member.move_to(new_voice)
+                        data["jtcc"][str(new_voice.id)] = {}
+                        data["jtcc"][str(new_voice.id)]["owner"] = member.id
+                        data["jtcc"][str(new_voice.id)]["channel_name"] = new_voice.name
+                        data["jtcc"][str(new_voice.id)]["channel_id"] = new_voice.id
+                        data["jtcc"][str(new_voice.id)]["state"] = "public"
+                        with open("json/join_to_create.json", "w") as json_file:
+                            json.dump(data, json_file, indent=4)
+                except AttributeError:
+                    print("NoneType")
+
+            elif len(before.channel.members) == 0:  # and after.channel is None:
+                if str(before.channel.id) in data["jtcc"]:
+                    await before.channel.delete()
+
+                    del data["jtcc"][str(before.channel.id)]
                     with open("json/join_to_create.json", "w") as json_file:
                         json.dump(data, json_file, indent=4)
-            except AttributeError:
-                print("NoneType")
-
-        elif len(before.channel.members) == 0:  # and after.channel is None:
-            if str(before.channel.id) in data["jtcc"]:
-                await before.channel.delete()
-
-                del data["jtcc"][str(before.channel.id)]
-                with open("json/join_to_create.json", "w") as json_file:
-                    json.dump(data, json_file, indent=4)
-            else:
-                return
+                else:
+                    return
+        else:
+            pass
 
 
 def setup(bot):
